@@ -12,7 +12,6 @@ import net.minecraft.entity.passive.SheepEntity;
 import net.minecraft.entity.projectile.FishingBobberEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.Items;
-import net.minecraft.util.ActionResult;
 import net.minecraft.util.Hand;
 import net.minecraft.util.collection.DefaultedList;
 import net.minecraft.util.hit.BlockHitResult;
@@ -62,28 +61,16 @@ public class TickListener {
                 return;
             }
             switch (AutoHarvest.instance.mode) {
-                case SEED:
-                    seedTick();
-                    break;
-                case HARVEST:
-                    harvestTick();
-                    break;
-                case PLANT:
-                    plantTick();
-                    break;
-                case Farmer:
+                case SEED -> seedTick();
+                case HARVEST -> harvestTick();
+                case PLANT -> plantTick();
+                case Farmer -> {
                     harvestTick();
                     plantTick();
-                    break;
-                case FEED:
-                    feedTick();
-                    break;
-                case FISHING:
-                    fishingTick();
-                    break;
-                case BONEMEALING:
-                    bonemealingTick();
-                    break;
+                }
+                case FEED -> feedTick();
+                case FISHING -> fishingTick();
+                case BONEMEALING -> bonemealingTick();
             }
             if (AutoHarvest.instance.mode != AutoHarvest.HarvestMode.FISHING)
                 AutoHarvest.instance.taskManager.Add_TickSkip(AutoHarvest.instance.configure.tickSkip.value);
@@ -106,6 +93,7 @@ public class TickListener {
                 for (int deltaZ = -configure.effect_radius.value; deltaZ <= configure.effect_radius.value; ++deltaZ) {
                     BlockPos pos = new BlockPos(X + deltaX, Y + deltaY, Z + deltaZ);
                     if (CropManager.isWeedBlock(w, pos) || (AutoHarvest.instance.configure.flowerISseed.value && CropManager.isFlowerBlock(w, pos))) {
+                        assert MinecraftClient.getInstance().interactionManager != null;
                         MinecraftClient.getInstance().interactionManager.attackBlock(pos, Direction.UP);
                         return;
                     }
@@ -128,10 +116,12 @@ public class TickListener {
                         if (b == Blocks.SWEET_BERRY_BUSH) {
                             BlockPos downPos = pos.down();
                             BlockHitResult blockHitResult = new BlockHitResult(new Vec3d(X + deltaX + 0.5, Y, Z + deltaZ + 0.5), Direction.UP, downPos, false);
-                            ActionResult a = MinecraftClient.getInstance().interactionManager.interactBlock(p, MinecraftClient.getInstance().world, Hand.MAIN_HAND, blockHitResult);
-                            String ass = "";
-                        } else
+                            assert MinecraftClient.getInstance().interactionManager != null;
+                            MinecraftClient.getInstance().interactionManager.interactBlock(p, MinecraftClient.getInstance().world, Hand.MAIN_HAND, blockHitResult);
+                        } else {
+                            assert MinecraftClient.getInstance().interactionManager != null;
                             MinecraftClient.getInstance().interactionManager.attackBlock(pos, Direction.UP);
+                        }
                         return;
                     }
                 }
@@ -153,13 +143,13 @@ public class TickListener {
         ItemStack itemStack = p.getMainHandStack();
         if (itemStack.isEmpty()) {
             if (lastUsedItem != null && !lastUsedItem.isEmpty()) {
-                DefaultedList<ItemStack> inv = p.inventory.main;
+                DefaultedList<ItemStack> inv = p.getInventory().main;
                 for (int idx = 0; idx < 36; ++idx) {
                     ItemStack s = inv.get(idx);
                     if (s.getItem() == lastUsedItem.getItem() &&
                             s.getDamage() == lastUsedItem.getDamage() &&
                             !s.hasTag()) {
-                        AutoHarvest.instance.taskManager.Add_MoveItem(idx, p.inventory.selectedSlot);
+                        AutoHarvest.instance.taskManager.Add_MoveItem(idx, p.getInventory().selectedSlot);
                         return s;
                     }
                 }
@@ -169,25 +159,27 @@ public class TickListener {
             return itemStack;
         }
     }
+
     /**
      * @return -1: does't have rod; 0: no change; change
      **/
     private int tryReplacingFishingRod() {
         ItemStack itemStack = p.getMainHandStack();
-        if (CropManager.isRod(itemStack) && (configure.keepFishingRodAlive.value == false || itemStack.getMaxDamage() - itemStack.getDamage() > 1)) {
+        if (CropManager.isRod(itemStack) && (!configure.keepFishingRodAlive.value || itemStack.getMaxDamage() - itemStack.getDamage() > 1)) {
             return 0;
         } else {
-            DefaultedList<ItemStack> inv = p.inventory.main;
+            DefaultedList<ItemStack> inv = p.getInventory().main;
             for (int idx = 0; idx < 36; ++idx) {
                 ItemStack s = inv.get(idx);
-                if (CropManager.isRod(s) && (configure.keepFishingRodAlive.value == false || s.getMaxDamage() - s.getDamage() > 1)) {
-                    AutoHarvest.instance.taskManager.Add_MoveItem(idx, p.inventory.selectedSlot);
+                if (CropManager.isRod(s) && (!configure.keepFishingRodAlive.value || s.getMaxDamage() - s.getDamage() > 1)) {
+                    AutoHarvest.instance.taskManager.Add_MoveItem(idx, p.getInventory().selectedSlot);
                     return 1;
                 }
             }
             return -1;
         }
     }
+
     private void plantTick() {
         ItemStack handItem = tryFillItemInHand();
         if (handItem == null) return;
@@ -212,7 +204,8 @@ public class TickListener {
                 w.getBlockState(pos.offset(Direction.DOWN)).canPlaceAt(w, pos);
                 if (CropManager.canPlantOn(handItem.getItem(), w, pos)) {
                     lastUsedItem = handItem.copy();
-                    ActionResult report = MinecraftClient.getInstance().interactionManager.interactBlock(MinecraftClient.getInstance().player, MinecraftClient.getInstance().world, Hand.MAIN_HAND, blockHitResult);
+                    assert MinecraftClient.getInstance().interactionManager != null;
+                    MinecraftClient.getInstance().interactionManager.interactBlock(MinecraftClient.getInstance().player, MinecraftClient.getInstance().world, Hand.MAIN_HAND, blockHitResult);
                     minusOneInHand();
                     return;
                 }
@@ -239,7 +232,8 @@ public class TickListener {
                         if (w.getBlockState(tmpPos).getBlock() == Blocks.AIR) {
                             lastUsedItem = handItem.copy();
                             BlockHitResult blockHitResult = new BlockHitResult(new Vec3d(X + deltaX + 1, Y + deltaY + 0.5, Z + deltaZ + 0.5), tmpFace, pos, false);
-                            ActionResult report = MinecraftClient.getInstance().interactionManager.interactBlock(MinecraftClient.getInstance().player, MinecraftClient.getInstance().world, Hand.MAIN_HAND, blockHitResult);
+                            assert MinecraftClient.getInstance().interactionManager != null;
+                            MinecraftClient.getInstance().interactionManager.interactBlock(MinecraftClient.getInstance().player, MinecraftClient.getInstance().world, Hand.MAIN_HAND, blockHitResult);
                             minusOneInHand();
                             return;
                         }
@@ -249,7 +243,8 @@ public class TickListener {
                         if (w.getBlockState(tmpPos).getBlock() == Blocks.AIR) {
                             lastUsedItem = handItem.copy();
                             BlockHitResult blockHitResult = new BlockHitResult(new Vec3d(X + deltaX, Y + deltaY + 0.5, Z + deltaZ + 0.5), tmpFace, pos, false);
-                            ActionResult report = MinecraftClient.getInstance().interactionManager.interactBlock(MinecraftClient.getInstance().player, MinecraftClient.getInstance().world, Hand.MAIN_HAND, blockHitResult);
+                            assert MinecraftClient.getInstance().interactionManager != null;
+                            MinecraftClient.getInstance().interactionManager.interactBlock(MinecraftClient.getInstance().player, MinecraftClient.getInstance().world, Hand.MAIN_HAND, blockHitResult);
                             minusOneInHand();
                             return;
                         }
@@ -259,7 +254,8 @@ public class TickListener {
                         if (w.getBlockState(tmpPos).getBlock() == Blocks.AIR) {
                             lastUsedItem = handItem.copy();
                             BlockHitResult blockHitResult = new BlockHitResult(new Vec3d(X + deltaX + 0.5, Y + deltaY + 0.5, Z + deltaZ + 1), tmpFace, pos, false);
-                            ActionResult report = MinecraftClient.getInstance().interactionManager.interactBlock(MinecraftClient.getInstance().player, MinecraftClient.getInstance().world, Hand.MAIN_HAND, blockHitResult);
+                            assert MinecraftClient.getInstance().interactionManager != null;
+                            MinecraftClient.getInstance().interactionManager.interactBlock(MinecraftClient.getInstance().player, MinecraftClient.getInstance().world, Hand.MAIN_HAND, blockHitResult);
                             minusOneInHand();
                             return;
                         }
@@ -269,7 +265,8 @@ public class TickListener {
                         if (w.getBlockState(tmpPos).getBlock() == Blocks.AIR) {
                             lastUsedItem = handItem.copy();
                             BlockHitResult blockHitResult = new BlockHitResult(new Vec3d(X + deltaX + 0.5, Y + deltaY + 0.5, Z + deltaZ), tmpFace, pos, false);
-                            ActionResult report = MinecraftClient.getInstance().interactionManager.interactBlock(MinecraftClient.getInstance().player, MinecraftClient.getInstance().world, Hand.MAIN_HAND, blockHitResult);
+                            assert MinecraftClient.getInstance().interactionManager != null;
+                            MinecraftClient.getInstance().interactionManager.interactBlock(MinecraftClient.getInstance().player, MinecraftClient.getInstance().world, Hand.MAIN_HAND, blockHitResult);
                             minusOneInHand();
                             return;
                         }
@@ -291,29 +288,44 @@ public class TickListener {
         ItemStack handItem = tryFillItemInHand();
         if (handItem == null) return;
         Collection<Class<? extends AnimalEntity>> animalList = CropManager.FEED_MAP.get(handItem.getItem());
+        if (animalList.isEmpty()) return;
         Box box = new Box(p.getX() - configure.effect_radius.value, p.getY() - configure.effect_radius.value, p.getZ() - configure.effect_radius.value,
                 p.getX() + configure.effect_radius.value, p.getY() + configure.effect_radius.value, p.getZ() + configure.effect_radius.value);
-        for (Class<? extends AnimalEntity> type : animalList) {
-            for (AnimalEntity e : p.getEntityWorld().getEntitiesByClass(type, box, null)) {
-                if (e.getBreedingAge() >= 0 && !e.isInLove()) {
-                    lastUsedItem = handItem.copy();
-                    ActionResult result = MinecraftClient.getInstance().interactionManager
-                            .interactEntity(p, e, Hand.MAIN_HAND);
-                }
-            }
-        }
         if (handItem.getItem() == Items.SHEARS) {
-            for (SheepEntity e : p.getEntityWorld().getEntitiesByClass(SheepEntity.class, box, null)) {
-                if (!e.isBaby() && !e.isSheared()) {
+            for (SheepEntity e : p.getEntityWorld().getEntitiesByClass(
+                    SheepEntity.class,
+                    box,
+                    sheepEntity -> !sheepEntity.isBaby() && !sheepEntity.isSheared()
+            )) {
+                lastUsedItem = handItem.copy();
+                assert MinecraftClient.getInstance().interactionManager != null;
+                MinecraftClient.getInstance().interactionManager.interactEntity(p, e, Hand.MAIN_HAND);
+                return;
+            }
+        } else {
+            /*
+            Special handling for axolotls: the food is a single use item and after it is used, it is replaced with a water
+            bucket.  The interaction is resolved on the server - if the client doesn't match, the next animal to be
+            interacted with gets scooped up rather than fed.
+            */
+            for (Class<? extends AnimalEntity> type : animalList) {
+                for (AnimalEntity e : p.getEntityWorld().getEntitiesByClass(
+                        type,
+                        box,
+                        animalEntity -> animalEntity.getBreedingAge() >= 0 && !animalEntity.isInLove()
+                )) {
                     lastUsedItem = handItem.copy();
-                    MinecraftClient.getInstance().interactionManager.interactEntity(p, e, Hand.MAIN_HAND);
-                    return;
+
+                    assert MinecraftClient.getInstance().interactionManager != null;
+                    MinecraftClient.getInstance().interactionManager
+                            .interactEntity(p, e, Hand.MAIN_HAND);
                 }
             }
         }
     }
 
     private long getWorldTime() {
+        assert MinecraftClient.getInstance().world != null;
         return MinecraftClient.getInstance().world.getTime();
     }
 
@@ -333,6 +345,7 @@ public class TickListener {
                 /* Reel */
                 if (fishBitesAt == 0 && isFishBites(p)) {
                     fishBitesAt = getWorldTime();
+                    assert MinecraftClient.getInstance().interactionManager != null;
                     MinecraftClient.getInstance().interactionManager.interactItem(
                             p,
                             MinecraftClient.getInstance().world,
@@ -341,6 +354,7 @@ public class TickListener {
 
                 /* Cast */
                 if (fishBitesAt != 0 && fishBitesAt + 20 <= getWorldTime()) {
+                    assert MinecraftClient.getInstance().interactionManager != null;
                     MinecraftClient.getInstance().interactionManager.interactItem(
                             p,
                             MinecraftClient.getInstance().world,
@@ -349,7 +363,6 @@ public class TickListener {
                 }
                 break;
             case 1:
-                return;
         }
     }
 
@@ -376,8 +389,10 @@ public class TickListener {
                         //not Mature
                         if (!((CropBlock) block).isMature(blockState)) {
                             BlockHitResult blockHitResult = new BlockHitResult(new Vec3d(X + deltaX + 0.5, Y, Z + deltaZ + 0.5), Direction.UP, pos, false);
+                            assert handItem != null;
                             lastUsedItem = handItem.copy();
-                            ActionResult report = MinecraftClient.getInstance().interactionManager.interactBlock(MinecraftClient.getInstance().player, MinecraftClient.getInstance().world, Hand.MAIN_HAND, blockHitResult);
+                            assert MinecraftClient.getInstance().interactionManager != null;
+                            MinecraftClient.getInstance().interactionManager.interactBlock(MinecraftClient.getInstance().player, MinecraftClient.getInstance().world, Hand.MAIN_HAND, blockHitResult);
                             minusOneInHand();
                             return;
                         }
