@@ -1,7 +1,8 @@
 package com.flier268.autoharvest.Plugin;
 
-import io.github.prospector.modmenu.api.ConfigScreenFactory;
-import io.github.prospector.modmenu.api.ModMenuApi;
+
+import com.terraformersmc.modmenu.api.ConfigScreenFactory;
+import com.terraformersmc.modmenu.api.ModMenuApi;
 
 public class ModMenu implements ModMenuApi {
     @Override
